Re-organize dashboard code
