+++ dashboard/layouts/fig.py
@@ -1,0 +1,40 @@
+import plotly.graph_objects as go
+from .styles import COLORS
+
+
+def blank_fig(text=None):
+    """Creates a blank figure."""
+    fig = go.Figure(data=go.Scatter(x=[], y=[]))
+    fig.update_layout(
+        paper_bgcolor=COLORS['blue4'],
+        plot_bgcolor=COLORS['blue4'],
+        width=300,
+        height=300)
+
+    fig.update_xaxes(showgrid=False, showticklabels=False, zeroline=False)
+    fig.update_yaxes(showgrid=False, showticklabels=False, zeroline=False)
+
+    if text is not None:
+        fig.update_layout(
+            width=300,
+            height=300,
+            annotations=[
+                        {
+                            "text": text,
+                            "xref": "paper",
+                            "yref": "paper",
+                            "showarrow": False,
+                            "font": {
+                                "size": 14,
+                                "color": COLORS['blue1']
+                            },
+                            "valign": "top",
+                            "yanchor": "top",
+                            "xanchor": "center",
+                            "yshift": 60,
+                            "xshift": 10
+                        }
+                    ]
+            )
+
+    return fig
