+++ dashboard/layouts/styles.py
@@ -1,0 +1,10 @@
+COLORS = {
+    'white': '#FFFFFF',
+    'text': '#091D58',
+    'blue1' : '#063446', #dark blue
+    'blue2' : '#0e749b',
+    'blue3' : '#15b3f0',
+    'blue4' : '#E4F3F9', #light blue
+    'red1'  : '#FF0000',
+    'yellow1' : '#f0d515'
+}
